rag/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.13
aiosignal==1.4.0
altgraph==0.17.4
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
beautifulsoup4==4.13.3
cachetools==5.5.2
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
contourpy==1.3.1
cycler==0.12.1
dataclasses-json==0.6.7
decorator==5.2.1
deep-translator==1.11.4
defusedxml==0.7.1
dill==0.3.9
distro==1.9.0
docling==2.29.0
docling-core==2.25.0
docling-ibm-models==3.4.1
docling-parse==4.0.1
easyocr==1.7.2
et_xmlfile==2.0.0
fastapi==0.115.14
filelock==3.17.0
filetype==1.2.0
fonttools==4.55.6
fpdf==1.7.2
frozenlist==1.7.0
fsspec==2024.12.0
google-ai-generativelanguage==0.6.18
google-api-core==2.25.1
google-api-python-client==2.175.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-generativeai==0.8.5
googleapis-common-protos==1.70.0
greenlet==3.2.3
grpcio==1.73.1
grpcio-status==1.71.2
h11==0.16.0
httpcore==1.0.9
httplib2==0.22.0
httpx==0.28.1
httpx-sse==0.4.1
huggingface-hub==0.30.2
idna==3.10
imageio==2.37.0
imageio-ffmpeg==0.6.0
Jinja2==3.1.5
jira==3.8.0
jiter==0.10.0
jsonlines==3.1.0
jsonpatch==1.33
jsonpointer==3.0.0
jsonref==1.1.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kiwisolver==1.4.8
langchain==0.3.26
langchain-community==0.3.27
langchain-core==0.3.68
langchain-experimental==0.3.4
langchain-google-genai==2.1.6
langchain-text-splitters==0.3.8
langsmith==0.4.4
latex2mathml==3.77.0
lazy_loader==0.4
llvmlite==0.44.0
lxml==5.3.2
markdown-it-py==3.0.0
marko==2.1.3
MarkupSafe==3.0.2
marshmallow==3.26.1
matplotlib==3.10.0
mdurl==0.1.2
more-itertools==10.6.0
moviepy==2.1.2
mpire==2.10.2
mpmath==1.3.0
multidict==6.6.3
multiprocess==0.70.17
mypy_extensions==1.1.0
networkx==3.4.2
ninja==1.11.1.4
numba==0.61.0
numpy==1.26.4
oauthlib==3.2.2
openai==1.93.0
openai-whisper==20240930
opencv-python-headless==4.11.0.86
openpyxl==3.1.5
orjson==3.10.18
packaging==23.2
pandas==2.2.3
pefile==2023.2.7
pillow==10.4.0
pluggy==1.5.0
proglog==0.1.11
propcache==0.3.2
proto-plus==1.26.1
protobuf==5.29.5
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.2
pyclipper==1.3.0.post6
pycparser==2.22
pydantic==2.11.3
pydantic-settings==2.8.1
pydantic_core==2.33.1
Pygments==2.19.1
pyinstaller==6.12.0
pyinstaller-hooks-contrib==2025.2
pylatexenc==2.10
pyparsing==3.2.1
pypdfium2==4.30.1
python-bidi==0.6.6
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-dotenv==1.1.0
python-pptx==1.0.2
pytz==2025.2
pywin32==310
pywin32-ctypes==0.2.3
PyYAML==6.0.2
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rich==14.0.0
rpds-py==0.24.0
rsa==4.9.1
rtree==1.4.0
safetensors==0.5.3
scikit-image==0.25.2
scipy==1.15.2
semchunk==2.2.2
setuptools==75.8.0
shapely==2.1.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
sounddevice==0.5.1
soupsieve==2.6
SQLAlchemy==2.0.41
starlette==0.46.2
sympy==1.13.1
tabulate==0.9.0
tenacity==8.5.0
tifffile==2025.3.30
tiktoken==0.8.0
tokenizers==0.21.1
torch==2.6.0
torchvision==0.21.0
tqdm==4.67.1
transformers==4.51.1
typer==0.12.5
typing-inspect==0.9.0
typing-inspection==0.4.0
typing_extensions==4.12.2
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.3.0
uvicorn==0.35.0
XlsxWriter==3.2.2
yarl==1.20.1
zstandard==0.23.0
